pyproject.toml
[project]
name = "chess-vlm"
version = "0.1.0"
description = "A toy analysis of the visual reasoning priors of VLMs"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "chessboard-image>=1.1.3",
    "datasets>=4.2.0",
    "ipykernel>=6.30.1",
    "ipywidgets>=8.1.7",
    "vllm>=0.11.0",
    "python-chess>=1.999",
    "pandas>=2.0.0",
    "tqdm>=4.65.0",
    "cairosvg>=2.8.2",
    "matplotlib>=3.10.7",
    "click>=8.0.0",
    "openai>=1.0.0",
    "python-dotenv>=1.0.0",
    "anthropic>=0.69.0",
]

[dependency-groups]
dev = [
    "ipykernel>=6.30.1",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]
